Change sorting local variables types to long long

+++ lab_1/stl.cpp
@@ -15,8 +15,8 @@ void primes(std::vector<int>& container) {
 }
 
 void sort(std::vector<int>& container) {
-    size_t len = container.size();
-    int i, j, gap;
+    long long len = static_cast<long long>(container.size());
+    long long i, j, gap;
     for (gap = len / 2; gap > 0; gap = gap / 2)
         for (i = gap; i < len; i++)
             for (j = i - gap; j >= 0 && container[j] > container[j+gap]; j = j - gap)
